+++ src/bot.py
@@ -18,8 +18,6 @@ from discord.utils import find
 from ext.const import log, bot, get_prefix
 from ext.const import statusLs, LASTWRDFILE, STRFILE, SETFILE, slog, nlog, hint, logging, cmdHdlLogger, eventLogger, style
 from ext import excepts
-import special
-from modules.chat import chat
 load_dotenv()
 cogs = []
 for cog in os.listdir('cogs/'):
@@ -43,28 +41,6 @@ except json.JSONDecodeError:
 MODE = os.getenv('MODE')
 setattr(logging.Logger, 'hint', hint)
 settings = json.load(open(SETFILE))
-slog("Adding bot commands...")
-
-
-# ---------
-# background tasks
-async def status():
-    await bot.wait_until_ready()
-    while True:
-        try:
-            if not MODE or MODE == 'NORMAL':
-                activity = discord.Game(name=random.choice(statusLs))
-                await bot.change_presence(status=discord.Status.online, activity=activity)
-            elif MODE == 'DEBUG':
-                activity = discord.Activity(type=discord.ActivityType(
-                    3), name="windowsboy111 debugging me")
-                await bot.change_presence(status=discord.Status.idle, activity=activity)
-            elif MODE == 'FIX':
-                await bot.change_presence(status=discord.Status.dnd, activity=discord.Activity(type=discord.ActivityType(3), name="windowsboy111 fixing me"))
-            await asyncio.sleep(30)
-        except Exception:
-            pass
-
 
 async def update():
     global settings, stringTable, lastword
@@ -72,14 +48,16 @@ async def update():
         open(LASTWRDFILE, 'r')), json.load(open(STRFILE, 'r'))
     settings.update(json.load(open(SETFILE, 'r')))
     json.dump(settings, open(SETFILE, 'w'))
-
+# ---------
+# background tasks
 
 async def task_update():
     await update()
     await asyncio.sleep(60)
 
-bot.loop.create_task(status())
 bot.loop.create_task(task_update())
+
+slog("Adding bot commands...")
 
 
 @bot.command(name='reboot', aliases=['restart'], hidden=True)
@@ -118,50 +96,17 @@ async def cmd_update(ctx):
     await update()
 
 
-# ------
-# events
-@bot.event
-async def on_message(message: discord.Message):
-    global lastword
-    if await special.pre_on_message(message):
-        return
-    try:
-        lastword[f'g{message.guild.id}'][str(message.author.id)] = message.id
-    except Exception:
-        lastword[f'g{message.guild.id}'] = {message.author.id: message.id}
-    if not isinstance(message.channel, discord.DMChannel) and message.channel.name == 'merlin-chat' and not message.author.bot:
-        async with message.channel.typing():
-            await message.channel.send(chat.response(message.content))
-        return 0
-    await update()
-    if message.content.startswith(get_prefix(bot, message)):
-        msgtoSend = f'{message.author} has issued command: '
-        print(msgtoSend + style.green + message.content + style.reset)
-        cmdHdlLogger.info(msgtoSend + message.content)
-        try:
-            await log(message.channel.mention + ' ' + msgtoSend + '`' + message.content + '`', guild=message.channel.guild)
-        except AttributeError:
-            pass
-        try:
-            await bot.process_commands(message)
-            return 0
-        except discord.ext.commands.errors.CommandNotFound:
-            pass
-        except discord.errors.NotFound:
-            pass
-        except Exception:
-            await message.channel.send(f'{message.author.mention}, there was an error trying to execute that command! :(')
-            print(traceback.format_exc())
-    special.post_on_message(message)
-
-
 @bot.event
 async def on_ready():
     nlog(f'Logged in as {bot.user.name} - {bot.user.id} in {MODE} mode')
     nlog('Loading Extensions...')
+    slog('Loading ext.tasks...')
+    bot.load_extension("ext.tasks")
+    slog('Loading ext.cmdhdl...')
+    bot.load_extension("ext.cmdhdl")
     try:
         for cog in cogs:
-            slog(f'Loading {cog}...')
+            slog(f'Loading cogs.{cog}...')
             bot.load_extension('cogs.' + cog)
     except Exception:
         nlog("An error occurred during loading extension, treat bot start as a reconnect")
@@ -179,98 +124,6 @@ async def on_ready():
         await log('*RUNNING IN EMERGENCY **FIX** MODE!')
     nlog("Ready!")
     return 0
-
-
-@bot.event
-async def on_member_join(member: discord.Member):
-    eventLogger.info(f"{member} has joined {member.guild}")
-    print(f"{member} has joined {member.guild}")
-    return 0
-
-
-@bot.event
-async def on_guild_join(guild: discord.Guild):
-    sudoers = []
-    for name in [role.name for role in guild.owner.roles]:
-        if ['admin', 'mod', 'owner'] in name.lower():
-            sudoers.append(name)
-    f = json.load(open(SETFILE, 'r'))
-    f[f'g{guild.id}'] = {"prefix": ['/'],
-                         "cmdHdl": {"cmdNotFound": 0}, "sudoers": sudoers}
-    with open(SETFILE, 'w') as outfile:
-        json.dump(f, outfile)
-    return 0
-
-
-@bot.event
-async def on_command_error(ctx, error):
-    try:
-        raise error
-    except Exception:
-        # This tells the issuer that the command cannot be used in DM
-        if isinstance(error, commands.errors.NoPrivateMessage):
-            try:
-                await ctx.author.send(f':x::lock: {ctx.command} cannot be used in Private Messages.')
-                return 3
-            except discord.HTTPException:
-                return 3
-        # This prevents any commands with local handlers being handled here in on_command_error.
-        if hasattr(ctx.command, 'on_error'):
-            return
-
-        # This prevents any cogs with an overwritten cog_command_error being handled here.
-        if ctx.cog and ctx.cog._get_overridden_method(ctx.cog.cog_command_error) is not None:
-            return
-
-        # Anything in ignored will return and prevent anything happening.
-        if isinstance(error, commands.errors.CommandNotFound):
-            try:
-                if settings[f'g{ctx.guild.id}']['cmdHdl']['cmdNotFound']:
-                    await ctx.send(":interrobang: Welp, I've no idea. Command not found!")
-            except KeyError:
-                await ctx.send(":interrobang: :two: :x:\n<:err:740034702743830549> Command not found!\n<:warn:739838316374917171> something went wrong, please run `/settings`")
-            return 2
-        if isinstance(error, commands.MissingRequiredArgument) or isinstance(error, commands.BadArgument):
-            return await ctx.invoke(bot.get_command('help'), cmdName=ctx.command.qualified_name)
-
-        if isinstance(error, commands.errors.DisabledCommand):
-            return await ctx.send(embed=discord.Embed(
-                title=f':no_entry: {ctx.command} has been disabled.',
-                description=f':x: `{ctx.message.content}`',
-                color=0xff0000
-            ))
-
-        if isinstance(error, commands.errors.CommandInvokeError):
-            await ctx.send(embed=discord.Embed(
-                title='<:err:740034702743830549> uh oh. An exception has occurred during the execution of the command',
-                description=stringTable['CommandInvokeError'].format(
-                    content=ctx.message.content),
-                timestamp=datetime.utcnow(),
-                color=0xff0000
-            ))
-
-        if isinstance(error, commands.errors.NotOwner):
-            await ctx.send(stringTable['notOwner'])
-            return 6
-        if isinstance(error, commands.errors.ConversionError):
-            await ctx.send(
-                ':bangbang: Hey bud, seems like you tried to input some invalid type of arguments to the command call!\n'
-                'Either CoNsUlT a PsYcHiAtRiSt or check the usage. Please!')
-            return 4
-
-        if isinstance(error, commands.errors.BadArgument):
-            await ctx.send(
-                ':grey_question: Whoops. The discord special expression you have specified when issuing that command is invalid.'
-                ':mag: This error occurrs usually because of the bot fails to find the object.')
-            return 4
-        if isinstance(error, excepts.NotMod):
-            return await ctx.send(str(error))
-
-        # All other Errors not returned come here. And we can just print the default TraceBack.
-        await log(f'Ignoring exception in command {ctx.message.content}:' + '\n\n```' + str(traceback.format_exc()) + '\n```', guild=ctx.guild)
-        return 1
-
-slog("Adding bot commands...")
 
 
 def start(token=None, **kwargs):